Correction de tous les bugs

+++ app/Http/Controllers/ProductController.php
@@ -5,12 +5,13 @@ namespace App\Http\Controllers;
 use App\Product;
 use App\Category;
 use Cache;
+use Storage;
 
 use Illuminate\Http\Request;
 
 class ProductController extends Controller
 {
-    private $paginate = 6;
+    private $paginate = 10;
 
     /**
      * Display a listing of the resource.
@@ -19,14 +20,9 @@ class ProductController extends Controller
      */
     public function index()
     {
-        $key = 'home' . ( request()->page ?? '1' );
-        $minutes = 5 * 60;
-
         // pour nettoyer le cache on a la commande
         // php artisan cache:clear
-        $products = Cache::remember( $key , $minutes, function(){
-            return Product::with('category')->paginate($this->paginate);
-        });
+        $products = Product::with('category')->paginate($this->paginate);
 
         return view('back.product.index', [
             'products'      => $products,
@@ -56,49 +52,39 @@ class ProductController extends Controller
      */
     public function store(Request $request)
     {
-        // dump($request->all()); die;
-
         // pour les validations vous avez les rules
         $request->validate([
             'title'            => 'required',
             'description'      => 'required',
             'price'            => 'required',
-            'category'         => 'required|integer',
-            'size'             => 'required|in:46,48,50,52',
-            'picture'          => 'required',
+            'category_id'      => 'required|integer',
+            'size'             => 'required',
             'picture'          => 'required|image|max:3000',
             'status'           => 'required|in:published,unpublished',
-            'code'             => 'required|in:new,soldes',
+            'code'             => 'required|in:new,solde',
             'reference'        => 'required|integer',
         ]);
 
+        // Formater les tailles 
+        $newSizeValue           = '['.implode(',',$request->size).']';
+
+        // Modifier la valeur de la taille dans request
+        $requestData            = $request->all();
+        $requestData['size']    = $newSizeValue;
 
         // insérer les données en base il faut préciser cela dans les fillables
-        $product = Product::create($request->all());
-        // une fois le produit créé en base de données Laravel crée un objet produit
-        // la méthode category()->attach permet d'associer dans la relation many to many des auteurs pour ce livre
-        $product->category()->attach($request->category);
+        $product = Product::create($requestData);
 
         if ($request->file('picture')) {
-            // enregistre dans le store l'image et en même temps on récupère le
-            // nom de l'image créé par Laravel, nom sécurisé
-            // 1. Pas d'écrasement d'image avec le même nom.
-            // 2. Pas d'injection de script dans le nom de l'image.
-            $link = $request->file('picture')->store('');
-
-            // créer une resource dans la table pictures
-            $product->picture = $link;
+            $link               = $request->file('picture')->store('');
+            $product->url_image = $link;
+            $product->save();
         }
 
         Cache::flush();
 
         // with permet de créer un flash message enregistrer dans la classe Session clé/valeur :
-        /*
-        return redirect()->route('book.index')->with('message', [
-            'type' => 'alert-success',
-            'content' => 'success create book'
-        ]);
-        */
+        return redirect()->route('product.index');
     }
 
     /**
@@ -137,7 +123,42 @@ class ProductController extends Controller
      */
     public function update(Request $request, Product $product)
     {
-        //
+
+        // pour les validations vous avez les rules
+        $request->validate([
+            'title'            => 'required',
+            'description'      => 'required',
+            'price'            => 'required',
+            'category_id'      => 'required|integer',
+            'size'             => 'required',
+            'picture'          => 'image|max:3000',
+            'status'           => 'required|in:published,unpublished',
+            'code'             => 'required|in:new,solde',
+            'reference'        => 'required|integer',
+        ]);
+
+        // Formater les tailles 
+        $newSizeValue           = '['.implode(',',$request->size).']';
+
+        // Modifier la valeur de la taille dans request
+        $requestData            = $request->all();
+        $requestData['size']    = $newSizeValue;
+
+        // insérer les données en base il faut préciser cela dans les fillables
+        $product->update($requestData);
+
+        if ($request->delete_picture) {
+            $this->deletePicture($product);
+        }
+        if ($request->picture) {
+            $link               = $request->file('picture')->store('');
+            $product->url_image = $link;
+            $product->save();
+        }
+
+        Cache::flush();
+        
+        return redirect()->route('product.index');
     }
 
     /**
@@ -148,11 +169,47 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {
-        Storage::disk('local')->delete($product->url_image);
+        if ($product->url_image) {
+            // on supprime physiquement l'image
+            Storage::disk('local')->delete($product->url_image);
+        }
+
         $product->delete();
 
         Cache::flush();
 
         return redirect()->route('product.index');
+    }
+
+    /**
+     * Remove picture
+     * 
+     * @param  Product $product
+     * @return void
+     */
+    private function deletePicture($product):void
+    {
+        if ($product->url_image) {
+            Storage::disk('local')->delete($product->url_image);
+        }
+    }
+
+    /**
+     * upload image 
+     * 
+     * @param  Product $product
+     * @param  Request $request
+     * @return void
+     */
+    private function uploadPicture($product, $request):void{
+
+            $this->deletePicture($product);
+            // enregistre dans le store l'image et en même temps on récupère le
+            // nom de l'image créé par Laravel, nom sécurisé
+            // 1. Pas d'écrasement d'image avec le même nom.
+            // 2. Pas d'injection de script dans le nom de l'image.
+            $link               = $request->file('picture')->store('');
+            $product->url_image = $link;
+            $product->save();
     }
 }
